split context and provider
