+++ windwatts-ui/src/providers/UnitsContext.tsx
@@ -1,4 +1,4 @@
-import { createContext, useState, useEffect } from "react";
+import { createContext } from "react";
 import { StoredUnits } from "../types/Units";
 
 export interface Units {
@@ -8,12 +8,12 @@ export interface Units {
   updateUnits: (newValues: StoredUnits) => void;
 }
 
-const defaultUnitValues: StoredUnits = {
+export const defaultUnitValues: StoredUnits = {
   windspeed: "m/s",
   output: "kWh",
 };
 
-const defaultValues: Units = {
+export const defaultValues: Units = {
   units: defaultUnitValues,
   setUnits: () => {},
   updateUnit: () => {},
@@ -21,47 +21,3 @@ const defaultValues: Units = {
 };
 
 export const UnitsContext = createContext<Units>(defaultValues);
-
-export default function UnitsProvider({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  const [units, setUnits] = useState<StoredUnits>(() => {
-    try {
-      const stored = localStorage.getItem("units");
-      return stored ? JSON.parse(stored) : defaultUnitValues;
-    } catch (error) {
-      console.error("Error reading from localStorage:", error);
-      return defaultUnitValues;
-    }
-  });
-
-  useEffect(() => {
-    try {
-      localStorage.setItem("units", JSON.stringify(units));
-    } catch (error) {
-      console.error("Error writing to localStorage:", error);
-    }
-  }, [units]);
-
-  const updateUnit = (key: string, value: string) => {
-    setUnits((prev) => ({
-      ...prev,
-      [key]: value,
-    }));
-  };
-
-  const updateUnits = (newValues: StoredUnits) => {
-    setUnits((prev) => ({
-      ...prev,
-      ...newValues,
-    }));
-  };
-  
-  return (
-    <UnitsContext.Provider value={{ units, setUnits, updateUnit, updateUnits }}>
-      {children}
-    </UnitsContext.Provider>
-  );
-}
